Fix integer overflow panic in tests after 0.10.1 update
while the cause was hard to figure out, it's the same story as
a couple of the walls that have been hit: now that allocators
are pass by value we can't create them in initializers.  They
are reclaimed when the method returns and it leads to craziness.

The implication is that we can't pass a base allocator to an
object that then wraps with it's own arena and deinit method.
We also can't assert that the passed allocator is of type arena
because std.mem.Allocator holds an opaque pointer.  It's a bit
of a bummer, and a possible solution is what we do in ir.zig
testSetup/testTeardown but that solution is a bit unsavory